fix: wrong use of NextResponse
In Next you're meant to use the .json() when sending a json response
(obviously) but if you need to set cookies then you need to first instantiate
a new NextResponse with a stringified json response, then set the cookies on that
response and return it from the handler. Which is really annoying.